Implemented find_datanode_for_block to use free_bytes and xmits
Live DNs which do not already have a replica of the requested block and have
enough free space to hold the block are pushed onto a priority queue based on
num transmits. Then, the number of reqeusted datanodes (replication_factor
many) are popped from the priority queue and returned to the caller.